+++ client/src/logic/mod.rs
@@ -1,6 +1,9 @@
-mod logic;
+mod account;
+mod device;
+mod key;
+mod message;
 
-pub use logic::{
-    fetch_prekeys, process_messages, provision_device, publish_prekeys, register_account,
-    send_message,
-};
+pub use account::register_account;
+pub use device::provision_device;
+pub use key::{fetch_prekeys, publish_prekeys};
+pub use message::{process_messages, send_message};
